+++ tests/unit/test_parsers/test_parse.py
@@ -1,5 +1,6 @@
 from pathlib import Path
 
+import numpy as np
 import pandas as pd
 import pytest
 
@@ -9,7 +10,7 @@ from app.parsers import (CalibreXMLParser, HSSParser, JSONParser, SSFileParser, 
 
 @pytest.fixture
 def opcfield():
-    return OPCFieldReverse(0, 0, 15, 15, 1, 2, 0, 5, 'B', 2)
+    return OPCFieldReverse(0, 0, 15, 15, 2, 1, 'B', 2)
 
 
 def test_get_parser(test_files):
@@ -33,8 +34,8 @@ def test_opcfield_reverse_base_dataframe(opcfield):
 
 def test_opcfield_reverse_with_ap(opcfield):
     expected = pd.DataFrame.from_dict({
-        'B2': {'x': 0, 'y': 0, 'row': 2, 'col': 'B', 'name': 'B2', 'x_ap': 0, 'y_ap': 5},
-        'C2': {'x': 15, 'y': 0, 'row': 2, 'col': 'C', 'name': 'C2', 'x_ap': 0, 'y_ap': 5}
+        'B2': {'x': 0, 'y': 0, 'row': 2, 'col': 'B', 'name': 'B2', 'x_ap': np.nan, 'y_ap': np.nan},
+        'C2': {'x': 15, 'y': 0, 'row': 2, 'col': 'C', 'name': 'C2', 'x_ap': np.nan, 'y_ap': np.nan}
     }, orient='index')
     pd.testing.assert_frame_equal(opcfield.parse_data(), expected,
                                   check_like=True, obj='parse_data output')
@@ -49,8 +50,8 @@ def test_parse_data(opcfield):
         'row': [2, 2],
         'col': ['B', 'C'],
         'name': ['B2', 'C2'],
-        'x_ap': [0, 0],
-        'y_ap': [5, 5]
+        'x_ap': [np.nan, np.nan],
+        'y_ap': [np.nan, np.nan]
     })
     expected_data.index = expected_data['name']
     expected_data.index.name = None
@@ -59,7 +60,7 @@ def test_parse_data(opcfield):
     # Vérification que les colonnes 'x_ap' et 'y_ap' sont présentes dans le DataFrame
     print(data.index)
     print(expected_data.index)
-    assert 'x_ap' in data.columns, "'x_ap' column is missing in the DataFrame"
-    assert 'y_ap' in data.columns, "'y_ap' column is missing in the DataFrame"
+    assert 'x_ap' in data.columns
+    assert 'y_ap' in data.columns
     pd.testing.assert_frame_equal(data, expected_data)
     # pd.testing.assert_frame_equal(opcfield.parse_data(), expected_data)